+++ packages/extension/babel.config.js
@@ -10,6 +10,8 @@ module.exports = {
         pragma: 'h',
       },
     ],
-    ['@babel/plugin-proposal-class-properties'],
+    ['@babel/plugin-proposal-class-properties', {"loose": true}],
+    ["@babel/plugin-proposal-private-methods", { "loose": true }],
+    ["@babel/plugin-proposal-private-property-in-object", { "loose": true }]
   ],
 };